+++ packages/cli/src/lib/service/Service.ts
@@ -3,7 +3,7 @@ import fg from 'fast-glob';
 // import os from 'node:os';
 import { pinyin } from 'pinyin-pro';
 // import { Worker } from 'node:worker_threads';
-import path, {
+import {
   join,
   isAbsolute,
 } from 'path';
@@ -17,7 +17,7 @@ import {
   renderAllApi,
 } from '@api-helper/template';
 import micromatch from 'micromatch';
-import { merge, pick, castArray } from 'lodash';
+import { merge, pick, castArray, isEqual } from 'lodash';
 import { APIHelper } from '@api-helper/core/lib/types';
 import { uuid, formatDate } from '@api-helper/core/lib/utils/util';
 import { checkDocument } from '@api-helper/template/lib/render-all-api';
@@ -108,11 +108,12 @@ class Service{
     // 添加解析插件
     this.injectParserPlugins(configList);
 
+    let hasGen = false;
     for (let i = 0; i < configList.length; i++) {
       try {
         const config = configList[i];
         if (len > 1) {
-          logger.info(`———————————————————— \x1B[34m${this.locales.$t('正在处理').replace('%0', String(i + 1))}\x1B[0m ————————————————————`);
+          logger.info(`———————————————————— \x1B[34m${this.locales.$t('正在处理').replace('%0', String(i))}\x1B[0m ————————————————————`);
         }
         // 缺少输出路径跳过该项
         if (!config.outputPath && !config.outputFilePath) {
@@ -131,10 +132,11 @@ class Service{
         const codes = await this.genCode(config, chooseDocumentList);
 
         await this.output(config, codes);
+        hasGen = true;
       } catch {}
     }
 
-    this.setApiHelperCLIRunningData();
+    hasGen && this.setApiHelperCLIRunningData();
     await this.clear();
   }
 
@@ -221,23 +223,48 @@ class Service{
 ${temp.join('\n')}
 // ==/ApiHelperCLIRunningData==`;
       }
+      let notChange = isEqual(this.selectedDocumentEtagTemp, this.apiHelperCLIRunningData.selectedDocumentEtag);
+      const removeBreakLine = (str: string) => {
+        str = str ?? '';
+        while (str.endsWith('\n') || str.endsWith('\r')) {
+          str = str.replace(/[\n\r]$/, '');
+        }
+        return str;
+      }
+      if (notChange) {
+        if (this.apiHelperCLIRunningData.selectedDocumentEtag.length === 0) {
+          let content = fsExtra.readFileSync(this.configFileAbsolutePath).toString();
+          const apiHelperCLIRunningData = content.match(/\/\/\s==ApiHelperCLIRunningData==([\s\S]*)\/\/\s==\/ApiHelperCLIRunningData==/im);
+          if (apiHelperCLIRunningData) {
+            content = content.replace(/\/\/\s==ApiHelperCLIRunningData==([\s\S]*)\/\/\s==\/ApiHelperCLIRunningData==/im, '');
+            content = removeBreakLine(content);
+            content += '\n';
+            fsExtra.writeFileSync(this.configFileAbsolutePath, content, { encoding: 'utf-8' });
+          }
+        }
+        return;
+      }
       let content = fsExtra.readFileSync(this.configFileAbsolutePath).toString();
+      const val = {
+        ...this.apiHelperCLIRunningData,
+        selectedDocumentEtag: this.selectedDocumentEtagTemp,
+      };
       let temp = toComment({
         ...this.apiHelperCLIRunningData,
         selectedDocumentEtag: this.selectedDocumentEtagTemp,
       });
+      // 取消生成内容
+      if (!val.selectedDocumentEtag.length) {
+        temp = '';
+      }
       const apiHelperCLIRunningData = content.match(/\/\/\s==ApiHelperCLIRunningData==([\s\S]*)\/\/\s==\/ApiHelperCLIRunningData==/im);
       if (apiHelperCLIRunningData) {
         content = content.replace(/\/\/\s==ApiHelperCLIRunningData==([\s\S]*)\/\/\s==\/ApiHelperCLIRunningData==/im, temp);
+        content = removeBreakLine(content);
+        content += '\n';
       } else {
-        if (!(content.endsWith('\n\n') || content.endsWith('\r\r'))) {
-          if (content.endsWith('\n') || content.endsWith('\r')) {
-            temp = '\n' + temp;
-          } else {
-            temp = '\n\n' + temp;
-          }
-        }
-        content += temp + '\n';
+        content = removeBreakLine(content);
+        content = content + (temp ? '\n\n' + `${temp}\n` : '\n');
       }
       fsExtra.writeFileSync(this.configFileAbsolutePath, content, { encoding: 'utf-8' });
     } catch (e: any) {