+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # 更新日志
 
+## 3.3.2
+
+### Patch Changes
+
+- v3.3.2
+
+## 3.3.1
+
+### Patch Changes
+
+- v3.3.1
+
+## 3.3.0
+
+### Minor Changes
+
+- v3.3.0
+
 ## 3.2.0
 
 ### Minor Changes